Revert to "normal" pad sizes as float.
Will investigate later on if need be the gain (if any) the storage of the padsizes as int (in micrometers then) would bring.

+++ genmap/genPadSize.h
@@ -10,7 +10,7 @@ namespace mch
 {
 namespace mapping
 {
-using PadSizeArray = std::array<std::pair<int,int>,18>;
+using PadSizeArray = std::array<std::pair<float,float>,18>;
 extern PadSizeArray arrayOfPadSizes;
 
 } // namespace mapping
